+++ glace/glace-manager.h
@@ -31,7 +31,7 @@ struct _GlaceManagerClass {
 
     // methods
     void (*capture_client)(GlaceManager* self, GlaceClient* client, gboolean overlay_cursor, GlaceManagerCaptureClientCallback callback, gpointer user_data, GDestroyNotify notify);
-    void (*capture_client_handle)(GlaceManager* self, gint handle, gboolean overlay_cursor, GlaceManagerCaptureClientCallback callback, gpointer user_data, GDestroyNotify notify);
+    void (*capture_client_handle)(GlaceManager* self, gint64 handle, gboolean overlay_cursor, GlaceManagerCaptureClientCallback callback, gpointer user_data, GDestroyNotify notify);
 };
 
 struct _GlaceManagerPrivate {
@@ -53,7 +53,7 @@ enum {
 GType glace_manager_get_type();
 GlaceManager* glace_manager_new();
 void glace_manager_capture_client(GlaceManager* self, GlaceClient* client, gboolean overlay_cursor, GlaceManagerCaptureClientCallback callback, gpointer user_data, GDestroyNotify notify);
-void glace_manager_capture_client_handle(GlaceManager* self, gint handle, gboolean overlay_cursor, GlaceManagerCaptureClientCallback callback, gpointer user_data, GDestroyNotify notify);
+void glace_manager_capture_client_handle(GlaceManager* self, gint64 handle, gboolean overlay_cursor, GlaceManagerCaptureClientCallback callback, gpointer user_data, GDestroyNotify notify);
 G_END_DECLS
 
 #endif /* __LIBGLACE_MANAGER_H__ */
